feat: implement MINK explicit enum discriminants
`enum E { A = 5, B }` explicit discriminants on unit and data-carrying
variants, with implicit continuation (previous value plus one), the
wrapping 64-bit literal model, duplicate rejection (E-T31, with the
earlier variant as the related location) and implicit-continuation
overflow rejection (E-T32); tags flow through construction, pattern
matching, equality, tagged-union layout, and native execution. Tag
values widen to i64 across the pipeline. Full suite: 1039 tests.

+++ src/backend/emit/x86_64.rs
@@ -1395,7 +1395,7 @@ fn emit_inst(
             // (tag) word is written at its offset, then the variant's own
             // payload bytes are copied into the payload area.
             let word0 = slots[target.raw() as usize].0;
-            code.movabs_rax(*discriminant as i64);
+            code.movabs_rax(*discriminant);
             code.mov_rbp_rax(word0 - *tag_offset as i32);
             if let Some(payload) = payload {
                 let dst = word0 - *payload_offset as i32;
